requirements.txt
fastapi==0.104.1
motor==3.3.2
pydantic==2.5.0
bcrypt==4.1.1
pyjwt==2.8.0
uvicorn==0.24.0
python-dotenv==1.0.0
